Tweaked appearance and handling of timer in Punchy game

+++ src/main/java/it/dpg/minigames/punchygame/view/PunchygameView.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface PunchygameView extends MinigameView {
     void updateSacks(final List<Direction> sacks);
     void updateScore(final int score);
-    void updateTimer(final int timer);
+    void updateTimer(final float timer);
     void updateBoxer(final Direction boxerDirection);
     void setInputObserver(final InputObserver observer);
 }
